requirements.txt
flask==3.1.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
werkzeug==3.1.3
openai==1.35.14
autogen
groq
language_tool_python
pdfplumber
python-docx
python-magic
pymongo==4.7.3
streamlit
pandas
qdrant-client
sentence-transformers
plotly.express